requirements.txt
# Snowflake dependencies
snowflake-connector-python==3.15.0
snowflake-sqlalchemy==1.5.1
SQLAlchemy==2.0.27

# Shopify API
ShopifyAPI>=12.3.0
python-dotenv==1.0.1
PyYAML==6.0.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
lightgbm==4.3.0
scikit-learn==1.3.2

# API and Web Framework
fastapi==0.104.0
uvicorn==0.23.2

# Environment and Configuration
pydantic==2.4.2

# Testing
pytest>=7.0.0
pytest-cov==4.1.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
tqdm==4.66.1
tabulate==0.9.0
pytz==2024.1
certifi>=2024.2.2
urllib3>=2.2.0

# Development
black>=22.0.0
flake8>=6.0.0

# Additional packages
cryptography==41.0.7
typing-extensions==4.9.0 